Bots/requirements.txt
requests
progressbar2
discord.py
python-dotenv==1.0.0